ai_automation_using_gauge/requirements.txt
getgauge
flask>=3.0.0
python-dotenv>=1.0.0
groq>=0.9.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
numpy>=1.26.0
playwright>=1.44.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
getgauge>=0.3.0
jinja2>=3.1.0
Pillow>=10.3.0
requests>=2.32.0
pyyaml>=6.0.1